Refactor package structure and add default setup.py and __init__.py for packaging
