+++ shapleqclient/common/error.py
