+++ shapleqclient/common/exception.py
