+++ shapleqclient/consumer.py
@@ -1,12 +1,11 @@
 from dataclasses import dataclass
-import logging
-from client import ClientBase, QConfig
-from exception import InvalidMessageError, SocketClosedError
-from proto.data_pb2 import SessionType
-from proto.api_pb2 import FetchResponse, Ack, BatchFetchResponse
+from shapleqclient.base import ClientBase, QConfig
+from shapleqclient.common.exception import InvalidMessageError, SocketClosedError
+from shapleqclient.proto.data_pb2 import SessionType
+from shapleqclient.proto.api_pb2 import FetchResponse, Ack, BatchFetchResponse
 from typing import Generator, Iterable
-from message.qmessage import QMessage, MessageType, make_qmessage_from_proto
-from message.api import fetch_msg
+from shapleqclient.message.qmessage import QMessage, MessageType, make_qmessage_from_proto
+from shapleqclient.message.api import fetch_msg
 
 
 @dataclass
